// admin/src/app/_cargo-company/model/cargoCompany.ts
export class CargoCompany {
    id: number;
    company: string;
    officer: string;
    phone: string;
    mobil: string;
    email: string;
    _id: string;

}

export interface ICargoCompany {
    id: number;
    company: string;
    officer: string;
    phone: string;
    mobil: string;
    email: string;
    _id: string;
}